+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [D] Changes is `Cere Dev` Runtime
 
 ## [vNext]
+- DAC nodes can be registered
+- Node mode is bitwise-or instead of enum
 - Added ChargeError event to payout pallet
 
 ## [4.8.9]